BLD: Do SWIG as part of build_ext instead of in subprocess
Instead of calling a subprocess which does the swig file before
calling setup.py. We can have build_ext call SWIG if we make
setup call build_ext before it tries to copy any files. Also,
we need to tell setup.py that the root of the project is src.

https://stackoverflow.com/questions/12491328/python-distutils-not-include-the-swig-generated-module